requirements.txt
# Graph:
    neo4j==5.22.0
    pydantic==2.7.4
# frontend:
    streamlit==1.36.0
# NER:
    gliner
    torch
    transformers
# misc:
    python-dotenv==1.0.1
# LLM:
    openai==1.35.7
